add version file

+++ src/utils/utils.h
@@ -7,9 +7,9 @@
 
 #if defined(ANDROID)
 #define ntohll(n) \
-  ((((uint64_t)ntohl((uint32_t)n)) << 32) + ntohl((uint32_t)(n >> 32)))
+  ((((uint64_t)ntohl((uint64_t)n)) << 32) + ntohl((uint32_t)((uint64_t)n >> 32)))
 #define htonll(n) \
-  ((((uint64_t)htonl((uint32_t)n)) << 32) + htonl((uint32_t)(n >> 32)))
+  ((((uint64_t)htonl((uint64_t)n)) << 32) + htonl((uint32_t)((uint64_t)n >> 32)))
 #endif
 
 /// <summary>
